Move control flow stuff to shared
