--- apps/src/lib/control_flow/timeouts.rs
@@ -1,1 +1,0 @@
-pub use namada::types::control_flow::timeouts::*;
