+++ apps/src/lib/mod.rs
@@ -8,7 +8,6 @@
 pub mod cli;
 pub mod client;
 pub mod config;
-pub mod control_flow;
 pub mod logging;
 pub mod node;
 pub mod wallet;
